requirements.txt
psycopg2>=2.7,<3.0
astroid==2.3.1
Django==2.2.6
django-rest-framework==0.1.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2